+++ collector-exoplayer/src/test/java/com/bitmovin/analytics/exoplayer/ExoPlayerAdapterTest.kt
@@ -39,6 +39,7 @@ class ExoPlayerAdapterTest {
                     mockk(),
                     mockk(),
                     mockk(),
+                    mockk(),
                 ),
                 recordPrivateCalls = true,
             )