+++ collector-media3-exoplayer/src/test/java/com/bitmovin/analytics/media3/exoplayer/Media3ExoPlayerAdapterTest.kt
@@ -39,6 +39,7 @@ class Media3ExoPlayerAdapterTest {
                     mockk(),
                     mockk(),
                     mockk(),
+                    mockk(),
                 ),
                 recordPrivateCalls = true,
             )