+++ collector/src/androidTest/java/com/bitmovin/analytics/data/testutils/CollectorTestUtils.kt
@@ -8,6 +8,7 @@ import com.bitmovin.analytics.adapters.PlayerContext
 import com.bitmovin.analytics.api.AnalyticsConfig
 import com.bitmovin.analytics.api.DefaultMetadata
 import com.bitmovin.analytics.api.SourceMetadata
+import com.bitmovin.analytics.data.DeviceInformationProvider
 import com.bitmovin.analytics.data.EventData
 import com.bitmovin.analytics.data.PlayerInfo
 import com.bitmovin.analytics.enums.PlayerType
@@ -33,6 +34,7 @@ internal class DummyPlayerAdapter(
             analytics,
             playerContext,
             Looper.getMainLooper(),
+            deviceInformationProvider = DeviceInformationProvider(analytics.context),
         )
     override val isAutoplayEnabled: Boolean?
         get() = null